Add sorted drag and drop within columns

+++ priv/repo/migrations/20241105074234_create_candidates.exs
@@ -5,7 +5,7 @@ defmodule Wttj.Repo.Migrations.CreateCandidates do
     create table(:candidates) do
       add :email, :string, null: false
       add :status, :string, default: "new", null: false
-      add :position, :integer, null: false
+      add :position, :float, null: false
       add :job_id, references(:jobs, on_delete: :delete_all), null: false
 
       timestamps(type: :utc_datetime)
